Rename TencentAgent::Error to TencentError to avoid conflict
It has overridden ActiveModel::MassAssignmentSecurity::Error under
TencentAgent.
