+++ app/models/tencent_agent/tweets_gathering.rb
@@ -8,7 +8,7 @@ class TencentAgent
       list_ids.each do |list_id|
         begin
           gather_tweets_from_list(list_id, list_last_timestamp_map[list_id])
-        rescue Error => e
+        rescue TencentError => e
           error "Aborted tweets gathering: #{e.message}"
         rescue => e
           log_unexpected_error(e)
